4.21

+++ part4/blogs-backend/controllers/blogs.js
@@ -32,8 +32,28 @@ blogsRouter.post("/", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
+  }
+  const user = await User.findById(decodedToken.id);
+
+  if (!user) {
+    return response
+      .status(400)
+      .json({ error: "no users found in the database" });
+  }
+
   const id = request.params.id;
+  const blog = await Blog.findById(id);
+
+  if (blog.user.toString() !== user.id.toString())
+    return response
+      .status(401)
+      .json({ error: "cannot delete blogs created by other users" });
+
   await Blog.findByIdAndDelete(id);
+
   response.status(200).end();
 });
